Make local hermetic builds possible
How-to:

    cachi2 fetch-deps --dev-package-managers rpm
    cachi2 inject-files ./cachi2-output --for-output-dir /cachi2
    podman build -f Containerfile.centos-stream9 \
        --network none \
        --no-cache \
        --security-opt label=disable \
        --cap-add all \
        --device /dev/fuse

Note that it was necessary to remove two repos from the config file:
copr-fedora-bootc-fasttracks and baseos-devel. The centos-stream9 image,
at least currently, doesn't use any RPMs from those repos. The lockfile
therefore doesn't include any RPMs from either of them => cachi2 doesn't
include them in the cachi2.repo file and doesn't generate metadata for
them. But rpm-ostree starts by processing all the repos referenced in
the config and fails with

    error: Installing packages: Unknown rpm-md repository: baseos-devel (discovered 5)

Fix by referencing only repos that are actually needed.

Note2: it is necessary to set gpgcheck=0 for the appstream-devel repo.
The c9s-devel-compose.repo disables the gpgcheck for this repo as well,
but in the generated cachi2.repo, it is not disabled. Fix with a sed
command.

Signed-off-by: Adam Cmiel <acmiel@redhat.com>