Don't remove unused repos from config file
Instead, add unused RPMs to the lock file. This makes cachi2 generate
everything that's needed to resolve the issue described in the previous
commit. It comes at the cost of "false positives" in the SBOM (cachi2
reports every RPM in the lockfile, we intentionally added unused RPMs).

Not sure which approach is better.

Signed-off-by: Adam Cmiel <acmiel@redhat.com>